requirements.txt
# Core LLM & Framework
openai
langchain>=0.2.0
langchain-community

# Vector Store
chromadb

# Web framework
fastapi
uvicorn[standard]

# Data processing
pandas
tqdm

# Utils
python-dotenv
pydantic

# Dev/test
pytest

sentence_transformers